Refactor settings acquisition into separate class
