+++ src/ServicePulse/ConstantsFile.cs
@@ -4,19 +4,17 @@ using System.Reflection;
 
 class ConstantsFile
 {
-    public static string GetContent()
+    public static string GetContent(Settings settings)
     {
-        var defaultRoute = Environment.GetEnvironmentVariable("DEFAULT_ROUTE") ?? "/dashboard";
         var version = GetVersionInformation();
-        var showPendingRetry = Environment.GetEnvironmentVariable("SHOW_PENDING_RETRY") ?? "false";
 
         var constantsFile = $$"""
 window.defaultConfig = {
-  default_route: '{{defaultRoute}}',
+  default_route: '{{settings.DefaultRoute}}',
   version: '{{version}}',
   service_control_url: '/api/',
   monitoring_urls: ['/monitoring-api/'],
-  showPendingRetry: {{showPendingRetry}},
+  showPendingRetry: {{settings.ShowPendingRetry}},
 }
 """;
