+++ src/ServicePulse/Program.cs
@@ -4,7 +4,9 @@ using ServicePulse;
 
 var builder = WebApplication.CreateBuilder(args);
 
-var (routes, clusters) = ReverseProxy.GetConfiguration();
+var settings = Settings.GetFromEnvironmentVariables();
+
+var (routes, clusters) = ReverseProxy.GetConfiguration(settings);
 builder.Services.AddReverseProxy().LoadFromMemory(routes, clusters);
 
 var app = builder.Build();
@@ -20,7 +22,7 @@ app.UseStaticFiles(staticFileOptions);
 
 app.MapReverseProxy();
 
-var constantsFile = ConstantsFile.GetContent();
+var constantsFile = ConstantsFile.GetContent(settings);
 
 app.MapGet("/js/app.constants.js", (HttpContext context) =>
 {
