+++ src/ServicePulse/ReverseProxy.cs
@@ -1,30 +1,18 @@
 ﻿namespace ServicePulse;
 
-using System.Text.Json;
 using Yarp.ReverseProxy.Configuration;
 using Yarp.ReverseProxy.Transforms;
 
 static class ReverseProxy
 {
-    public static (List<RouteConfig> routes, List<ClusterConfig> clusters) GetConfiguration()
+    public static (List<RouteConfig> routes, List<ClusterConfig> clusters) GetConfiguration(Settings settings)
     {
-        var serviceControlUrl = Environment.GetEnvironmentVariable("SERVICECONTROL_URL") ?? "http://localhost:33333";
-        var serviceControlUri = new Uri(serviceControlUrl);
-
-        var monitoringUrls = ParseLegacyMonitoringValue(Environment.GetEnvironmentVariable("MONITORING_URLS"));
-        var monitoringUrl = Environment.GetEnvironmentVariable("MONITORING_URL");
-
-        monitoringUrl ??= monitoringUrls;
-        monitoringUrl ??= "http://localhost:33633";
-
-        var monitoringUri = new Uri(monitoringUrl);
-
         var serviceControlInstance = new ClusterConfig
         {
             ClusterId = "serviceControlInstance",
             Destinations = new Dictionary<string, DestinationConfig>
             {
-                { "instance", new DestinationConfig { Address = serviceControlUri.GetLeftPart(UriPartial.Authority) } }
+                { "instance", new DestinationConfig { Address = settings.ServiceControlUri.GetLeftPart(UriPartial.Authority) } }
             }
         };
 
@@ -33,7 +21,7 @@ static class ReverseProxy
             ClusterId = "monitoringInstance",
             Destinations = new Dictionary<string, DestinationConfig>
             {
-                { "instance", new DestinationConfig { Address = monitoringUri.GetLeftPart(UriPartial.Authority) } }
+                { "instance", new DestinationConfig { Address = settings.MonitoringUri.GetLeftPart(UriPartial.Authority) } }
             }
         };
 
@@ -70,41 +58,5 @@ static class ReverseProxy
         };
 
         return (routes, clusters);
-    }
-
-    static string? ParseLegacyMonitoringValue(string? value)
-    {
-        if (value is null)
-        {
-            return null;
-        }
-
-        var cleanedValue = value.Replace('\'', '"');
-        var json = $$"""{"Addresses":{{cleanedValue}}}""";
-
-        MonitoringUrls? result;
-
-        try
-        {
-            result = JsonSerializer.Deserialize<MonitoringUrls>(json);
-        }
-        catch (JsonException)
-        {
-            return null;
-        }
-
-        var addresses = result?.Addresses;
-
-        if (addresses is not null && addresses.Length > 0)
-        {
-            return addresses[0];
-        }
-
-        return null;
-    }
-
-    class MonitoringUrls
-    {
-        public string[] Addresses { get; set; } = [];
     }
 }
